+++ types/signable.go
@@ -11,12 +11,12 @@ import (
 // Signable is an interface for all signable things.
 // It typically removes signatures before serializing.
 type Signable interface {
-	WriteSignBytes(chainID string, w io.Writer, n *int64, err *error)
+	WriteSignBytes(chainID string, w io.Writer, n *int, err *error)
 }
 
 // SignBytes is a convenience method for getting the bytes to sign of a Signable.
 func SignBytes(chainID string, o Signable) []byte {
-	buf, n, err := new(bytes.Buffer), new(int64), new(error)
+	buf, n, err := new(bytes.Buffer), new(int), new(error)
 	o.WriteSignBytes(chainID, buf, n, err)
 	if *err != nil {
 		PanicCrisis(err)
